requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
async-generator==1.10
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.10
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
